old auth lead to error scene - fixed

+++ src/page/TodoPage.js
@@ -1,19 +1,20 @@
-import { useEffect, useState } from 'react';
+import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../context/AuthContext';
 import TodoNav from '../component/TodoNav';
 import TodoAdd from '../component/TodoAdd';
 import List from '../component/List';
 
+const user = JSON.parse(window.localStorage.getItem('user'));
+
 function TodoPage() {
   const navigate = useNavigate();
   const [list, setList] = useState([]);
   const { token, setToken } = useAuth();
-  const user = JSON.parse(window.localStorage.getItem('user'));
 
   const getList = async () => {
-    console.log("get");
-      await fetch('https://todoo.5xcamp.us/todos', {
+    console.log('get');
+    await fetch('https://todoo.5xcamp.us/todos', {
       method: 'GET',
       headers: new Headers({
         'Content-Type': 'application/json',
@@ -21,14 +22,21 @@ function TodoPage() {
       }),
     })
       .then((res) => res.json())
-      .then((result) => setList(result.todos))
-      .catch((err) => console.log(err.toString()))
+      .then((result) => {
+        if (result.message === '未授權') {
+          navigate('../');
+          window.localStorage.clear();
+        } else {
+          setList(result.todos);
+        }
+      })
+      .catch((err) => console.log(err.toString()));
   };
 
-  const handleLogOut = (e) => {
+  const handleLogOut = async (e) => {
     e.preventDefault();
+    await window.localStorage.clear();
     setToken(null);
-    window.localStorage.clear();
     navigate('/');
   };
 
@@ -39,14 +47,14 @@ function TodoPage() {
       getList();
     }
   }, []);
-  
+
   return (
     <div id="todoListPage" className="bg-half">
       <TodoNav nickname={user?.nickname} handleLogOut={handleLogOut} />
       <div className="conatiner todoListPage vhContainer">
         <div className="todoList_Content">
           <TodoAdd getList={getList} />
-          <List list={list} setList={setList} getList={getList} />
+          <List list={list} getList={getList} />
         </div>
       </div>
     </div>
